refactor: 重命名生成器项目
